LayoutTests/http/tests/security/cross-frame-access-object-getPrototypeOf-expected.txt

CONSOLE MESSAGE: line 1: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 10: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
This tests that you can't get the prototype of the window or history objects cross-origin using Object.getPrototypeOf().

PASS: Object.getPrototypeOf(targetWindow) should be 'undefined' and is.
PASS: targetWindow.history should be 'undefined' and is.
